# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "BrowseRequest": {
        "properties": {
          "query": {
            "description": "Anki search to run in the Browser",
            "title": "Query",
            "type": "string"
          },
          "reorder": {
            "description": "{\"columnId\": ..., \"order\": \"ascending\"|\"descending\"}",
            "title": "Reorder"
          }
        },
        "title": "BrowseRequest",
        "type": "object"
      },
      "BrowseResult": {
        "properties": {
          "card_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Card Ids",
            "type": "array"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          }
        },
        "title": "BrowseResult",
        "type": "object"
      },
...
      "CurrentCard": {
        "description": "Null when no review is in progress.",
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "buttons": {
            "items": {
              "type": "integer"
            },
            "title": "Buttons",
            "type": "array"
          },
          "card_id": {
            "title": "Card Id",
            "type": "integer"
          },
          "css": {
            "title": "Css",
            "type": "string"
          },
          "deck_name": {
            "title": "Deck Name",
            "type": "string"
          },
          "field_order": {
            "title": "Field Order",
            "type": "integer"
          },
          "fields": {
            "additionalProperties": {
              "type": "object"
            },
            "title": "Fields",
            "type": "object"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "next_reviews": {
            "items": {
              "type": "string"
            },
            "title": "Next Reviews",
            "type": "array"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "template": {
            "title": "Template",
            "type": "string"
          }
        },
        "required": [
          "card_id",
          "fields",
          "field_order",
          "question",
          "answer",
          "buttons",
          "model_name",
          "deck_name",
          "css",
          "template"
        ],
        "title": "CurrentCard",
        "type": "object"
      },
      "CurrentCardResult": {
        "properties": {
          "card": {
            "$ref": "#/components/schemas/CurrentCard"
          },
          "review_active": {
            "default": false,
            "title": "Review Active",
            "type": "boolean"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          }
        },
        "title": "CurrentCardResult",
        "type": "object"
      },
...
      "GuiResult": {
        "description": "Did the UI do the thing.\n\nFalse rather than an error for the ordinary \"not right now\" cases - no\nBrowser open, no review in progress, no such deck - because those are\nstates a client polls for, not mistakes it made.",
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          }
        },
        "title": "GuiResult",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NoteIdList": {
        "properties": {
          "note_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Note Ids",
            "type": "array"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          }
        },
        "title": "NoteIdList",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "description": "Required when an API key is configured in Tsunagi settings.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "BearerAuth": {
        "description": "Alternative to the X-API-Key header.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "license": {
      "name": "MIT"
    },
    "title": "Tsunagi"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/gui/current-card": {
      "get": {
        "description": "`card` is null and `review_active` false when no review is in progress.",
        "operationId": "guiCurrentCard",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentCardResult"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "The card being reviewed",
        "tags": [
          "GUI"
        ],
        "x-permission": "gui"
      }
    },
    "/v1/gui/selected-notes": {
      "get": {
        "description": "Empty when no Browser is open.",
        "operationId": "guiSelectedNotes",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteIdList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Notes selected in the Browser",
        "tags": [
          "GUI"
        ],
        "x-permission": "gui"
      }
    },
...
    "/v1/gui:browse": {
      "post": {
        "description": "Opens the card Browser, optionally running a search and sorting it. Returns the card ids the query matches.",
        "operationId": "guiBrowse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrowseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowseResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Open the Browser",
        "tags": [
          "GUI"
        ],
        "x-permission": "gui"
      }
    },
...
    "/v1/gui:exit": {
      "post": {
        "description": "Shuts Anki down. Deferred by a second so this reply reaches you first.",
        "operationId": "guiExit",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuiResult"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Close Anki",
        "tags": [
          "GUI"
        ],
        "x-permission": "manage"
      }
    },
...
    "/v1/gui:undo": {
      "post": {
        "description": "Undoes the last operation, as Ctrl+Z would.",
        "operationId": "guiUndo",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuiResult"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Undo",
        "tags": [
          "GUI"
        ],
        "x-permission": "write"
      }
    },
...
  },
...
}